About section is now called Skills section
And other fixes were implemented
