+++ backend_sanity/schemas/skills.js
@@ -1,6 +1,6 @@
 export default{
-    name:'abouts',
-    title:'Abouts',
+    name:'skills',
+    title:'Skills',
     type: 'document',
     fields:[
         {
