+++ frontend_react/src/container/Skills/Skills.jsx
@@ -1,109 +1,58 @@
 import React, { useState, useEffect } from 'react';
 import { motion } from 'framer-motion';
-// import ReactTooltip from 'react-tooltip';
-import { Tooltip } from 'react-tooltip';
-import 'react-tooltip/dist/react-tooltip.css';
 
+// import { images } from '../../constants';
 import { AppWrap, MotionWrap } from '../../wrapper';
-// import { AppWrap } from '../../wrapper';
-
-import { urlFor, client } from '../../client';
+// import AppWrap from '../../wrapper/AppWrap';
 import './Skills.scss';
+import { urlFor, client } from '../../client';
 
+// const abouts = [
+//   { title: 'Web Development', description: ' I am a good web developer', imgUrl: images.about01 },
+//   { title: 'Web Design', description: 'I am a good web designer', imgUrl: images.about02 },
+//   { title: 'UI/UX', description: 'I am a good web developer', imgUrl: images.about03 },
+//   { title: 'Web Animations', description: 'I am a good web animator', imgUrl: images.about04 }
+// ];
 
-
-const Skills = () => {
-  const [experiences, setExperiences] = useState([]);
+const Skill = () => {
   const [skills, setSkills] = useState([]);
 
   useEffect(() => {
-    const query = '*[_type == "experiences"]';
-    const skillsQuery = '*[_type == "skills"]';
+    const query = '*[_type == "skills"]';
 
     client.fetch(query).then((data) => {
-      setExperiences(data);
-    });
-
-    client.fetch(skillsQuery).then((data) => {
       setSkills(data);
     });
   }, []);
 
   return (
     <>
-      <h2 className="head-text">Skills & Experiences</h2>
+      <h2 className="head-text"><span>Technologies</span> of expertise</h2>
 
-      <div className="app__skills-container">
-        <motion.div className="app__skills-list">
-          {skills.map((skill) => (
-            <motion.div
-              whileInView={{ opacity: [0, 1] }}
-              transition={{ duration: 0.5 }}
-              className="app__skills-item app__flex"
-              key={skill.name}
-            >
-              <div
-                className="app__flex"
-                style={{ backgroundColor: skill.bgColor }}
-              >
-                <img src={urlFor(skill.icon)} alt={skill.name} />
-              </div>
-              <p className="p-text">{skill.name}</p>
-            </motion.div>
-          ))}
-        </motion.div>
-        <div className="app__skills-exp">
-          {experiences.map((experience) => (
-            <motion.div
-              className="app__skills-exp-item"
-              key={experience.year}
-            >
-              <div className="app__skills-exp-year">
-                <p className="bold-text">{experience.year}</p>
-              </div>
-              <motion.div className="app__skills-exp-works">
-                {experience.works.map((work) => (
-                  
-                  <>
+      <div className="app__profiles">
+        {skills.map((skill, index) => (
+          <motion.div
+            whileInView={{ opacity: 1 }}
+            whileHover={{ scale: 1.1 }}
+            transition={{ duration: 0.5, type: 'tween' }}
+            className="app__profile-item"
+            key={skill.title + index}
+          >
+            <img src={urlFor(skill.imgUrl)} alt={skill.title} />
 
-                    <motion.div
-                      whileInView={{ opacity: [0, 1] }}
-                      transition={{ duration: 0.5 }}
-                      className="app__skills-exp-work"
-                      data-tip
-                      data-for={work.name}
-                      key={work.name}
-                       data-tooltip-content={work.desc}
-                      id="custom-inline-styles" 
-                    >
-                      <h4 className="bold-text">{work.name}</h4>
-                      <p className="p-text">{work.company}</p>
-                    </motion.div>
-
-                    <Tooltip
-                      anchorId="custom-inline-styles"
-                      // id={work.name}
-                      // effect="solid"
-                      // arrowColor="#000"
-                      className="skills-tooltip"
-                      // content={work.desc}
-                    >
-                    </Tooltip>
-                                      
-                  </>
-                ))}
-              </motion.div>
-            </motion.div>
-          ))}
-        </div>
+            <h2 className="bold-text" style={{ marginTop: 20 }}>{skill.title}</h2>
+            <p className="p-text" style={{ marginTop: 10 }}>{skill.description}</p>
+          </motion.div>
+        ))};
       </div>
     </>
   );
 };
 
-// export default AppWrap(Skills, 'skills')
+// export default About
+// export default AppWrap(About, 'about');
 export default AppWrap(
-  MotionWrap(Skills, 'app__skills'),
-  'skills',
-  'app__whitebg',
+  MotionWrap(Skill, 'app__skill'),
+  'skill',
+  // 'app__whitebg',
 );
