+++ frontend_react/src/container/Skills/Skills.scss
@@ -1,162 +1,53 @@
-.app__skills {
+.app__skill {
     flex: 1;
     width: 100%;
     flex-direction: column;
+    margin-top: 2rem;
   }
   
-  .app__skills-container {
-    width: 80%;
-    margin-top: 3rem;
-  
+  .app__profiles {
     display: flex;
-    flex-direction: row;
-  
-    @media screen and (max-width: 900px) {
-      width: 100%;
-      flex-direction: column;
-    }
-  }
-  
-  .app__skills-list {
-    flex: 1;
-    display: flex;
+    justify-content: center;
+    align-items: flex-start;
     flex-wrap: wrap;
+    margin-top: 2rem;
+    // background-color: whitesmoke;
+
+
+  }
+  
+  .app__profile-item {
+    width: 190px;
+    display: flex;
     justify-content: flex-start;
     align-items: flex-start;
-  
-    margin-right: 5rem;
-  
-    @media screen and (max-width: 900px) {
-      margin-right: 0;
-      justify-content: center;
-      align-items: center;
-    }
-  }
-  
-  .app__skills-item {
     flex-direction: column;
-    text-align: center;
+    margin: 2rem;
   
-    margin: 1rem;
+
+
+
+    img {
+      width: 100%;
+      height: 170px;
+      border-radius: 15px;
+      object-fit: cover;
+
+
+// About skill boxes
+// margin: 4rem;
+padding: 1rem;
+border-radius: 0.5rem;
+background-color: rgba(255, 255, 255, 0.6);
+color: #000;
+    }
   
-    transition: all 0.3s ease-in-out;
-  
-    div {
-      width: 90px;
-      height: 90px;
-      border-radius: 50%;
-      background-color: #fef4f5;
+    @media screen and (min-width: 2000px) {
+      width: 370px;
+      margin: 2rem 4rem;
   
       img {
-        width: 50%;
-        height: 50%;
+        height: 320px;
       }
-  
-      &:hover {
-        box-shadow: 0 0 25px #fef4f5;
-      }
-  
-      @media screen and (min-width: 2000px) {
-        width: 150px;
-        height: 150px;
-      }
-  
-      @media screen and (max-width: 450px) {
-        width: 70px;
-        height: 70px;
-      }
-    }
-  
-    p {
-      font-weight: 500;
-      margin-top: 0.5rem;
-    }
-  
-    @media screen and (min-width: 2000px) {
-      margin: 1rem 2rem;
-  
-      p {
-        margin-top: 1rem;
-      }
-    }
-  }
-  
-  .app__skills-exp {
-    flex: 1;
-    display: flex;
-    justify-content: flex-start;
-    align-items: flex-start;
-    flex-direction: column;
-  
-    @media screen and (max-width: 900px) {
-      margin-top: 2rem;
-    }
-  }
-  
-  .app__skills-exp-item {
-    width: 100%;
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-start;
-    align-items: flex-start;
-    margin: 1rem 0;
-  }
-  
-  .app__skills-exp-works {
-    flex: 1;
-  
-    .app__skills-exp-work {
-      display: flex;
-      flex-direction: column;
-      justify-content: flex-start;
-      align-items: flex-start;
-      margin-bottom: 1rem;
-      cursor: pointer;
-  
-      h4 {
-        font-weight: 500;
-      }
-  
-      p {
-        font-weight: 400;
-        color: var(--gray-color);
-        margin-top: 5px;
-      }
-
-
-      
-    }
-  }
-  
-  .app__skills-exp-year {
-    margin-right: 3rem;
-  
-    p {
-      font-weight: 800;
-      color: var(--secondary-color);
-    }
-  
-    @media screen and (max-width: 450px) {
-      margin-right: 1rem;
-    }
-  }
-  
-  .skills-tooltip {
-    max-width: 300px ;
-    background-color: var(--white-color) ;
-    box-shadow: 0 0 25px rgba(0, 0, 0, 0.1) ;
-    border-radius: 5px ;
-    padding: 1rem ;
-    color: var(--gray-color) ;
-    text-align: center ;
-    line-height: 1.5 ;
-    opacity: 1 ;
-    // background-color: aqua;
-    
-  
-    @media screen and (min-width: 2000px) {
-      font-size: 1.75rem ;
-      max-width: 500px ;
-      line-height: 2 ;
     }
   }
