+++ frontend_react/src/container/index.js
@@ -1,4 +1,4 @@
-export { default as About } from './About/About';
+// export { default as About } from './About2/About2';
 export { default as Footer } from './Footer/Footer';
 export { default as Header } from './Header/Header';
 export { default as Skills } from './Skills/Skills';
